[NEW RBO] Add PhysicalAggregationBuilder (#31657)
